pyproject.toml
[tool.poetry]
name = "refractiveindex"
version = "0.1.0"
description = ""
authors = ["stuermer <stuermer@lsw.uni-heidelberg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8, <3.12"
pyyaml = "^6.0"
requests = "^2.28.2"
scipy = "^1.10.1"
numba = "^0.56.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
